vo_vdpau: support display sync
Just needs to accept a NULL argument to draw_image() for frame repeats.
Unfortunately, it seems we can't just reuse the current output surface,
because an output surface can be queued only once at the same time to a
presentation queue. It might still be possible to make this more
efficient.